Reupload of a previous push I had done because there was version control issues. This should have changes last made by Ono at approximately 7 pm on 230404. The motor scaling changes should be additional to what was already done. If not, we need to fix this.
Fixed motor stall torque scaling. Previously, it took the original database's motor mass and used it for scaling. This led to a constant value. Now, the scaling law uses the new motor mass which has been scaled with the rest of the body. It also divides the scaled motor mass by 4 to find the mass of one motor and uses this value in the scaling calculation. This returns more reasonable values.